Increase ContractExecutionTransaction::SIGNED_VBYTES constant
We were only testing against CETs with a single output, so the
`SIGNED_VBYTES` constant was based on a smaller transaction than
usual.

We could use a different constant for the case when the CET has a
single output, but it's just an edge case and the transaction fees are
very small.