Remove snapshot tests
These tests don't really add much value anymore.
They were initially added for making sure we don't break the API towards the UI, but we introduced a separate model for that in the meanwhile.
One could still argue that these tests ensure backwards compatibility for the cfd state in the DB, but that is not important at the moment.
Since we have more complex states now it is quite a pain to maintain these tests now - without getting much out.