Add meta_state to cfd for simplifying ui
On a granular level, a CFD can be in multiple states. On an abstract level a CFD is either opening (not on chain yet), open (on chain) or closed (it has been terminated).
By introducing this meta state into our cfd model we simplify the UI logic.

Additionally, the daemon decides the color scheme for position and state badges.