Replace `.unwrap` with direct mapping to our `Error` type
When we want to map from one error type to the other, we can either
rely on the conversion via `From` that happens through the `?` operator
or map the `E` type of `Result` directly via the `map_err` function.

The latter is useful if we want to map to a specific variant like
"CannotInvertMatrix" in our case which captures the exact thing
that failed. Often times, libraries have a single `Error` type for
improved ergonomics and hence, simply converting every instance of that
error to our error via `From` would lose certain details as to which
operation precisely failed.

We still annotate our inner error with `#[source]` to instruct the
`thiserror` macro to render the error string of the inner error when
it is being printed out.