Transition cfd to new state when handling an event
Previously we only returned the new state, but we also have to set it internally, otherwise the cfd might be out of date.

This was a problem when getting the refund transaction from the cfd directly after a state transition to `MustRefund`.
The cfd instance had not actually transitioned to the new state, even though it was saved in the database.