Remove the spread
It does not make sense to put a spread on top of the price; this just results in a shitty UX, because the user always starts with a loss.
We need proper interest on top of the amount, i.e. a fixed rate that the maker always gets that is independent of the price.