Cleanup `awaiting_status` in case we don't have remaining subscriptions
We always replaced it with `remaining`, but when populating `awaiting_status` upon sync we only go by the keys, so we actually fill in the script status again.
We should remove the complete entry if we want to make sure that we don't want to process it anymore.

Note: Kept seeing monitoring logs for things that were already past Finality (...) so we looked into this.