Use the original order's term for rolling over
This allows us to avoid the use of the constant in two places which
removes the dependency on the constant from the taker entirely.

The logic here is generally not great - in particular having the taker
anticipate, which announcement it is going to be - but this is the
least invasive way of making the term dynamic.