Encrypt network communication using the noise protocol
The which do not fit into the 65535 byte message size
limit of the noise protocol are chunked before
encryption. This feature was introduced to handle the
very large messages which contained the CET's.

More research is required into the NOISE_PARAMS and
whether they can be used to configure snow elegantly
handle authentication.