Move settlement oracle event id into the Dlc
Oracle event-id and Dlc have to exist together, one without the other is incomplete, hence we move the event id inside.
Note that the ID is actually already know  to the Dlc as part of the `cet` hashmap, but this was not cleaned up in this patch.

This patch goes towards getting the model boundaries right to transition the architecture towards an event model.