Limit scope of the database connections acquired during application setup
The dirty shutdown we observed were caused by the sqlite pool patiently waiting
until we close the connection we've opened in the application setup.
This is turn was eventually triggering Rocket timeout (which is by default
configured to wait 5s for cleanup).

Tighten the scope of connections needed during application startup only to the
necessary minimum (housekeeping & loading cfds at startup)