Remember UTXOs previously used for creating transactions
This should prevent accidential double-spends of our own inputs
when doing concurrent contract setups or other interactions of the
wallet.

We only store these in-memory, thus a simple restart of the application
will "unlock" a UTXO in case it was previously marked used but the
transaction never made it to the blockchain for whatever reason.

Resolves #790.