Make all fields on CFD private
Making all fields private revealed some weirdness in how we compute
the position of the CFD. For an order, it made sense that we store
the `origin` of it. For a CFD however, it makes more sense to store
the position and role of the actor explicitly and derive other
information from there.