Don't manually lock EncryptedJsonCodec
There is a utility in `futures` for splitting anything that implements
`AsyncRead` + `AsyncWrite` into two components. It also uses a lock
interally but at least we don't have to implement it ourselves. In this
case, it also fixes using the wrong mutex (std::sync). A side-effect of
this change is that our codec now needs to have two type parameters
because we only construct it once and we have two different messages for
sending and receiving. The implications of this are minimal though.

Might help with #759.