Rewrite docker container and binary release process
Our requirements for releasing artefacts are:

1. Publish a multi-arch docker image on every push to master
2. Publish a multi-arch docker image for every release
3. Attach binaries for all platforms for every release

We want to do all of this:

- Without too much duplication
- As idiomatic as possible
- As performant as possible

The new solutions strives to achieve this.

To reduce duplication, we extract a GitHub action for building a
release binary for a particular platform. This action can be reused
between the workflow for building the docker image and the workflow
for attaching the binaries to a release.

Building multi-arch container images is a bit of a trickery. One
can use `docker buildx` but that unfortunately builds all containers
in one command. This is not ideal because we cannot leverage GitHub's
matrix to parallelise the building of the binaries. Fortunately, there
is a second way of building multi-arch images: `docker manifest`.

Each docker image has a manifest. A multi-arch docker image is simply
a manifest that refers to other manifests. We can leverage this by
building each platform specific docker image within a matrix build and
later, once all these jobs are completed, create a single manifest image
that bundles all the other images together.

To not overly complicate the docker publishing workflow, we separate the
two requirements of building an image for pushes to master and releases
into two workflows. By creating a reuseable GitHub workflow, we can
parameterise the docker tag that changes between those two workflows.