Don't use spawn_fallible for the rollover task
The only error that can occur here is "Actor is disconnected". That can
only happen if we discard the rollover actor before the protocol
finishes. In that case, we are simply no longer interested in the result
and we can thus safely ignore this error.

To make sure that this thinking is not invalidated, we add an explicit
type annotation that enforces that we are not dropping an error here in
case someone ever changes the return value from the handler to NOT be a
`()` Ok value.