Model `maker_inc_connections::Actor` as a resource around the listener
This actor can only operate for as long as the listener is active.
If the listener fails, we cannot do our job anymore so we need to shutdown.

Currently, no restart is implemented but we are also currently not handling
this so this is a pure refactoring and not a degradation in features.

Because we are no hiding the listener inside the actor system, the tests can
no longer grab the actual listening port and hand it to the taker. To circumvent
this, we temporarily bind to the 0 port, which allocates us a free port by the
kernel and pass in an explicit listening address.

This is in-line with ADR 006.