Align heartbeat variables in the taker
Both maker and taker actor systems are now constructed with the same heartbeat
interval variable.

Taker internally configures the heartbeat timeout as 2 * heartbeat_interval.
The heartbeat pulse (rate of checking whether timeout lapsed) is configured as
heartbeat_interval / 2. Beforehand, it was set to the same value as heartbeat
timeout, which could result in reporting a timeout even if it should not do so.

Adds more comments to describe all variables in use for future readers.