Remove `max_giveable`
This function was a horrendous hack introduced in the xmr-btc codebase
for figuring out how much we can maximally withdraw from a wallet.

Introducing it here is unnecessary because we can utilize BDK native
functionality to drain the entire wallet as we don't need to know
exactly, how much we are withdrawing if we are draining it.