Invert price-feed and projection dependencies
The price-feed is the more standalone component, hence it makes more
sense for the projection actor to depend on the price-feed rather than
the other way round.

This introduces a "polling" communication pattern but at the same time
simplifies the price-feed and will allow us to further embed it in the
actor system.