Embed price-feed actor in actor system
This allows us to access the latest quote from the price-feed via the
`LatestQuote` message directly within `propose_settlement`. This removes
the hack of converting the projected quote back into a
`bitmex_price_feed::Quote` within the taker's routes. It also increases
the encapsulation of our actor system and reduces setup boilerplate
within the binary.

To properly mock the price feed, we need to write the mock ourselves because
mockall doesn't allow us to _overwrite_ an expectation. That is, if we start
out with a mock of `None` for the current price, we cannot later override it
with `Some(quote)`. Writing the mock by hand allows us to do this and is not
actually much more code.