Run CPU expensive operations on a dedicated thread
Runtimes such as tokio have worker threads for executing futures. When a
futures yields control back to the runtime because it cannot do any more
work (i.e. it returns `Poll::Pending` from its `poll` function), a
different future is scheduled on the worker thread by the runtime.

If a future executes CPU intensive code (such as signature verification
/ creation), its `poll` function does not return at all up until this is
complete, thus occupying the worker thread until it is finished. In a
production setup, this is usually not a problem because runtimes like
tokio operate their worker threads in a work-stealing fashion which
means that an idle worker thread will take futures that are ready, steal
them from the original worker thread's queue and poll them.

This does become a problem in situations where there is only a single
worker thread. By default, `#[tokio::test]` creates a runtime with only
a single worker thread. In such situations, futures that would be ready
to execute are not polled until the CPU-intensive operation completes
which can lead to all kinds of weird behaviour.

To fix this, we instruct tokio to execute the CPU-intensive operations
of signing and verifying on separate a thread, dedicated for blocking
code.