Have `assert_next_state` wait until the given state
Previously, the macro would assert that the state of the CFD on the
next state change in the feed is a certain state. This can make the
tests brittle if there is another update in the feed inbetween where
the state is not yet the expected one.

Instead of asserting on the next feed update, we rewrite
`assert_next_state` as `wait_next_state` and spin in a loop until the
CFD satisfies a given predicate which in the case of the macro simply
performs an equality check on the state of the CFD.

This might potentially fix #1064.