Increase timeout to 180 seconds
When running all tests locally, cargo by default executes as many
tests as you have cores. That is 8 for me. On 8 cores, only 8 threads
can meaningfully run concurrently. For each test that sets up a contract
tokio ends up spawning 1 additional thread for blocking CPU work per
party (maker & taker). We have more than 4 tests which means not all
of these threads can utilize the CPU 100% for their own work. This makes
the entire contract setup take quite a bit longer because the kernel
ends up scheduling all of these threads to get some work done.

In the meantime, the timeout value which we are changing here is ticking
inside the test because it is waiting for contract setup to finish. We
need to increase this to account for this potential delay.