Only emit commit if we haven't done so already
As soon as the commit transaction hits the blockchain, the timelock for
using a CET starts ticking. If the oracle attests before the timelock
expires, the CET is still locked and hence, we cannot use it yet.

Previously, we assumed that the commit transaction hasn't been published
yet but that is not completely true.

Commit might have already gone on chain but the timelock simply hasn't
expired yet. In this case, we shouldn't emit the commit transaction to
allow consumers of our event to reason about this situation. In
particular, the projection for the UI doesn't need to add the commit
transaction URL in that case.