Move quantity state into `Trade` component
This state is not needed further up, we can easily move it into the
`Trade` component which reduces the amount of props we need to pass.

We drop the notion of an `effectiveQuantity` in favor of updating
the currently stored quantity conditionally. That should be easier
to maintain because we cannot accidentally mix up `quantity` and
`effectiveQuantity`.