Include transactions in more cases
Previously, we only included links to transactions if they were part of
an event. This turned out to not be true. Some transactions, like commit
can be published by either party in which case we only learn about it as
part of getting the `CommitConfirmed` event.

Refactor projection to explicitly track the status of whether a
transaction has been published for all transactions that we can derive
from the DLC. In case they have been published, compute the transaction
URL from the DLC and include it in the response.