Don't do async work in monitor constructor
This should speed up startup of the actor systems.

To achieve this, we need to move a few things around and introduce a new
message containing all the required params for re-initializing
monitoring. Moving re-broadcasting to its own task allows us to inline
the `broadcast` function again because we now need to send a message to
ourselves anyway as we don't have access to the actor's state within the
task.