Upgrade tracing-subscriber to 0.3
This version replaces `chrono` with `time` due to `chrono`'s dependency
on an older version of time that is vulnerable to segfaults [0]. The fix
for the vulnerability is to not allow multi-threaded access to the
timezone configured on the local system. Because our program is
multi-threaded, we can no longer use the local time for displaying the
logs unless we explicitely opt in to this potentially unsound behaviour
with the `unsound_local_offset` cfg. I don't want to think through
whether or not this is safe in our circumstances and hence opting to
change the logger to use UTC time instead.

[0]: https://nvd.nist.gov/vuln/detail/CVE-2020-26235