Populate minimal opening fee in the projection
Unfortunately, a Cfd offer from the maker is not enough to adequately calculate
opening fees, as we need quantity to do that.

For the time being, opt for calculating the fees for the min_quantity.
The total expected fees can be easily estimated by the user from that (or the UI
could even do a simple multiplication by the parcel size.

Another simpler approach could be the UI displaying:
    "opening fees from xxx for min_quantity"