Populate next funding event in the projection
Move the field from order to cfd - we don't know the next funding event in the
order, as it hasn't been taken yet.

Estimate the value based on the next full hour from the latest contract setup or
rollover completion.

In fullness of time, these events should probably communicate when the next
rollover will happen.