Correct payout and p/l API
The profit and loss should always be based on the payout.
If we know the payout because we either know the collaborative settlement tx or the cet, then we should use the transaction to determine the payout and the profit and loss based on that payout.
Of we don't have a spending transaction we determine the payout of the current price and then calculate the profit on top of that.