Configure global lints in .cargo/config
In the absence of a way to configure lints across a workspace, we use
a technique described in https://github.com/EmbarkStudios/rust-ecosystem/pull/68.

This allows us to move the lints that should unconditionally apply
to all crates to this one config, unblocking the split up of the mega
crate into multiple smaller ones.