Don't implement `FromParam` for `CfdAction`
If we want to decouple the rocket dependency from our core crate, we
cannot have this implementation. This looks like a step backwards but
actually, we should solve this by not parsing these actions at all and
instead send links to the clients.

See https://github.com/itchysats/itchysats/issues/955.