Manually implement `MockMonitor`
In order to have a more elaborate `MockMonitor` that can be instructed
to trigger commands on a CFD, we need to manually implement the mock.

This turns out to be pretty easy because all functions on
`monitor::Actor` return `()` or `Result<()>` and we always mocked them
in the same way.