Interact with monitor mock in tests
Instead of interacting with the ``{taker,maker}_cfd::Actor` via
`deliver_event!`, we interact with the monitor mock directly. This makes
more sense because we are mocking monitor in the tests. As such, we
should also trigger the behaviour that we want from the mocked system
through the mocked system.

Furthermore, this will allow us to remove the indirection of the monitor
actor to forward events to `{taker,maker}_cfd::Actor`, aligning the
monitor actor with ADR013.