Avoid log spam for wire messages
In case an xtra message handler hangs, it logs the message that has
been processing. For messages containing types from within the `wire`
module, these can be huge. So big that processing the logs becomes
impossible and/or exceeds certain limits of downstream software.

We already have specific Display implementations for these types, use
those for the Debug representation.