Add supervisor::Actor to xtras
The `supervisor::Actor` is a pattern generally applicable to actor
frameworks, so it doesn't apper to belong in `daemon`.

Unfortunately, we need to introduce a dependency to `xtras` in
`maker`. Maybe the `MakerActorSystem` can be reworked to remove this
dependency, or perhaps the dependency is fine.

Fixes https://github.com/itchysats/itchysats/issues/952.