Reduce built-in rocket logs
The built-in startup logs are noisy and don't add much value. Plus, it
is nice to control the log message we are seeing.

Eventually, it would be good to make rocket an implementation detail
such that a user doesn't directly know, which web framework we used to
implement the application.

Additionally, we disable the CLI colours to make the logs more readable
in downstream applications such as grafana.

Fixes #1278.