Cache expensive CFD events to avoid deserialising them from JSON
Examining the maker with a flamegraph showed that we have been using ~50% CPU on
deserialising CFD events from JSON.

Using `cached` crate reduced that number significantly (to ~8%). However, the
cache size constraint would become a problem for long-running CFDs - to minimise
this issue, selective dispatch for expensive events was used.

ContractSetupCompleted and RolloverCompleted store the whole DLC, which takes a
lot of time to deserialise; these events were chosen to be cached.