Only the taker publishes `commit` if collab settlement failed/rejected
The maker has no incentive to publish commit because if the cfd remains open periodic fees are collected.

With this change we can more confidently transition into a failed state if we don't have a connection to the collaborative settlement actor upon accept/reject for the maker.
If we would always publish commit in such an error scenario that could have weird side effect.