Add timeout to the taker's collab settlement setup actor
At the moment the taker waits indefinitely on the accept/reject of the maker.
Similar to the implementation for rollover we return failed in case we have not gotten an answer from the maker within 30 seconds.
Since we don't have a message channel in this protocol actor we record of the maker replied in a boolean value upon receiving the reply.