Treat RevokeConfirmed as CommitConfirmed
Fixes #1300

The problem was that we monitor for the publishment of revoked commit transaction but don't act on it yet. If there is a `todo!` in the code, we will panic and cannot recover from this.

While if the other party publishes a revoked commit transaction we should punish them this logic is not yet implemented.

For now, we just pretend that everything is fine and treat it as if it was the expected commit transaction.
The downside is that we will receive our money according to an old price.